+++ spec/rubocop/cop/rails/helper_instance_variable_spec.rb
@@ -36,4 +36,28 @@ RSpec.describe RuboCop::Cop::Rails::HelperInstanceVariable do
       end
     RUBY
   end
+
+  it 'does not register an offense when a class which inherits `ActionView::Helpers::FormBuilder`' do
+    expect_no_offenses(<<~'RUBY')
+      class MyFormBuilder < ActionView::Helpers::FormBuilder
+        def do_something
+          @template
+          @template = do_something
+        end
+      end
+    RUBY
+  end
+
+  it 'registers an offense when using a class which does not inherit `ActionView::Helpers::FormBuilder`' do
+    expect_offense(<<~'RUBY')
+      class Foo
+        def do_something
+          @template
+          ^^^^^^^^^ Do not use instance variables in helpers.
+          @template = do_something
+          ^^^^^^^^^ Do not use instance variables in helpers.
+        end
+      end
+    RUBY
+  end
 end